cast su Divisione.java

+++ src/codice/Divisione.java
@@ -1,3 +1,3 @@
 public class Divisione {
-    public static float divisione(int a,int b){return a/b;}
+    public static float divisione(int a,int b){return (float) a /b;}
 }
